refactor(avatar): creating avatar provider to prevent multiple settings calls
