+++ packages/avatar/Avatar.d.ts
@@ -1,13 +1,19 @@
-type SkeletonType = {
-    width?: string | number;
-    height?: string | number;
-};
-
-export interface AvatarProps {
-    fallback?: string;
-    skeletonProps?: SkeletonType; 
+export interface AvatarContext {
+  loading: boolean;
+  avatar?: string;
 }
 
+export interface AvatarProps {
+  children?:
+    | React.ReactNode
+    | ((avatarContext: AvatarContext) => React.ReactNode);
+  fallback?: string;
+}
+
+declare function useAvatarContext(): AvatarContext;
+
 declare const Avatar: React.FunctionComponent<AvatarProps>;
+
+export { useAvatarContext };
 
 export default Avatar;
