+++ packages/avatar/AvatarImage.d.ts
@@ -1,0 +1,12 @@
+type SkeletonType = {
+  width?: string | number;
+  height?: string | number;
+};
+
+export interface AvatarImageProps {
+  skeletonProps?: SkeletonType;
+}
+
+declare const AvatarImage: React.FunctionComponent<AvatarImageProps>;
+
+export default AvatarImage;
