+++ packages/avatar/index.d.ts
@@ -1,2 +1,7 @@
-export { default } from './Avatar';
-export * from './Avatar';
+import Avatar, { useAvatarContext, AvatarContext } from './Avatar';
+import AvatarImage from './AvatarImage';
+
+export default Avatar;
+
+export { AvatarImage, useAvatarContext, AvatarContext };
+
