+++ packages/avatar/index.js
@@ -1,3 +1,6 @@
-import Avatar from './Avatar';
+import Avatar, { useAvatarContext, AvatarContext } from './Avatar';
+import AvatarImage from './AvatarImage';
 
 export default Avatar;
+
+export { AvatarImage, useAvatarContext, AvatarContext };
